Chris_solution/project3/requirements.txt
Flask
pybase64
jsonpickle
Pillow
numpy
opencv-python
fastapi
tflite
tensorflow
tensorflow-hub
tflite-runtime
requests